* master:
  fix warning C4996: 'strcpy': This function or variable may be unsafe. Consider using strcpy_s instead.
  atlmfc: cleanup
  fix x64 warnings
  fix x64 warnings
  fix x64 warnings
  minor fix: use wchar_t
  OpenSessionInPutty: sync with WinSCP
  fix cppcheck warning: Member variable is not initialized in the constructor
  fix cppcheck warning: Invalid number of character ({) when these macros are defined: 'MPEXT_NO_SSL'
  fix cppcheck warning: The scope of the variable can be reduced
  fix cppcheck warning: Possible null pointer dereference: MTime - otherwise it is redundant to check it against null
  fix cppcheck warning: Using 'memset' on class that contains a 'std::basic_string
  fix cppcheck warning: Member variable 'TReadDirectoryProgressAction::ResolvedLinks' is not initialized in the constructor
  fix cppcheck warning: Assignment of function parameter has no effect outside the function
  fix includes
  refactoring: cleanup includes
  refactoring: use #pragma once
  fix code style
  use #pragma once

+++ src/core/Common.cpp
@@ -1850,7 +1850,7 @@ UnicodeString FixedLenDateTimeFormat(const UnicodeString & Format)
   return Result;
 }
 //---------------------------------------------------------------------------
-static UnicodeString FormatTimeZone(long /* Sec */)
+static UnicodeString FormatTimeZone(intptr_t /* Sec */)
 {
   // TTimeSpan Span = TTimeSpan::FromSeconds(Sec);
   UnicodeString Str;
@@ -2123,18 +2123,18 @@ uintptr_t ContinueAnswer(uintptr_t Answers)
   return Result;
 }
 //---------------------------------------------------------------------------
-UnicodeString LoadStr(int Ident, intptr_t MaxLength)
+UnicodeString LoadStr(intptr_t Ident, intptr_t MaxLength)
 {
   UnicodeString Result;
   Result.SetLength(MaxLength > 0 ? MaxLength : 1024);
   HINSTANCE hInstance =  GetGlobalFunctions()->GetInstanceHandle();
   assert(hInstance != 0);
-  intptr_t Length = static_cast<intptr_t>(::LoadString(hInstance, Ident, reinterpret_cast<LPWSTR>(const_cast<wchar_t *>(Result.c_str())), (int)Result.Length()));
+  intptr_t Length = static_cast<intptr_t>(::LoadString(hInstance, (UINT)Ident, reinterpret_cast<LPWSTR>(const_cast<wchar_t *>(Result.c_str())), (int)Result.Length()));
   Result.SetLength(Length);
   return Result;
 }
 //---------------------------------------------------------------------------
-UnicodeString LoadStrPart(int Ident, int Part)
+UnicodeString LoadStrPart(intptr_t Ident, intptr_t Part)
 {
   UnicodeString Result;
   UnicodeString Str = LoadStr(Ident);
